requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-jalali-date==1.0.2
django-js-asset==2.1.0
jdatetime==4.1.1
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
